refactor MaterialFarmerOption from struct to GroupOption
